+++ src/managers/ProfileManager.ts
@@ -1,5 +1,5 @@
 import fetch from "isomorphic-fetch"
-import { ManagerOptions } from "../types/types"
+import { ManagerOptions, WrapperOptions } from "../types/types"
 import { Profile, Report, ApiID } from "fagc-api-types"
 import BaseManager from "./BaseManager"
 import strictUriEncode from "strict-uri-encode"
@@ -7,10 +7,10 @@ import strictUriEncode from "strict-uri-encode"
 export default class ProfileManager extends BaseManager<Report> {
 	public apikey?: string
 	private apiurl: string
-	constructor(apiurl: string, apikey?: string, options: ManagerOptions = {}) {
-		super(options)
-		if (apikey) this.apikey = apikey
-		this.apiurl = apiurl
+	constructor(options: WrapperOptions, managerOptions: ManagerOptions = {}) {
+		super(managerOptions)
+		if (options.apikey) this.apikey = options.apikey
+		this.apiurl = options.apiurl
 	}
 	async fetchCommunity(playername: string, communityId: ApiID): Promise<Profile> {
 		const fetched = await fetch(`${this.apiurl}/profiles/fetchcommunity/${strictUriEncode(playername)}/${strictUriEncode(communityId)}`)