+++ src/managers/RevocationManager.ts
@@ -1,5 +1,5 @@
 import fetch from "isomorphic-fetch"
-import { ManagerOptions } from "../types/types"
+import { ManagerOptions, WrapperOptions } from "../types/types"
 import { Revocation, ApiID } from "fagc-api-types"
 import BaseManager from "./BaseManager"
 import strictUriEncode from "strict-uri-encode"
@@ -7,10 +7,10 @@ import strictUriEncode from "strict-uri-encode"
 export default class RevocationManager extends BaseManager<Revocation> {
 	public apikey?: string
 	private apiurl: string
-	constructor(apiurl: string, apikey?: string, options: ManagerOptions = {}) {
-		super(options)
-		if (apikey) this.apikey = apikey
-		this.apiurl = apiurl
+	constructor(options: WrapperOptions, managerOptions: ManagerOptions = {}) {
+		super(managerOptions)
+		if (options.apikey) this.apikey = options.apikey
+		this.apiurl = options.apiurl
 	}
 	resolveID(revocationid: ApiID): Revocation|null {
 		const cached = this.cache.get(revocationid)