+++ changelog.md
@@ -1,3 +1,7 @@
+## UNRELEASED
+### Additions
+- Added Master API rules part to create and remove rules
+
 ## [1.9.1] - [2021-08-12]
 ### Bugfixed
 - Fixed that `Report.reportedTime` was not a date but said it was